Move core exchange dispatcher into core internal as well as unit tests
